Add a unique_list function
The unique_list function takes a list as parameter, and returns
a list without any duplicated elements. The order of the elements
is maintained (i.e. if the first occurrence of an element appears in
the list before the first occurrence of another, the former will
precede the latter in the final list).