types: aquilon: add new structure_entitlements in the system parameters
A new entitlement system is being added to the Aquilon broker. It will
allow to set entitlements of a given type (for instance, 'login' or
'root') and enable them by host, cluster, personality, archetype or
Grn/EonID.

These new entitlements need to be stored to be picked up by the
templates in order to act on them, and this is why the
structure_entitlements is being added: we will now have in the
structure_system_aquilon structure an "entitlements" dictionary that
will use the entitlements type as key, and a structure_entitlements
as value. In there, we will find entitlements per-EonID or per-user.
In the later case, the user type (also being added to the broker)
will also be mentioned (as to, for instance, insure some activities
are only being entitled for system users).