Add auto as a valid filesystem type
Extend the validation of the type resource in structure_filesystem
to allow "auto" to be used. This is useful for fstab entries which have
been created outside Quattor, and the fstype may not be known.

According to mount(8) the order in which fs types are tried can be
controlled using /etc/filesystems.